fix(presentation): casting issue fixed in schadule task time
    - in LibraryApp class, cast the scheduler object passed time value
    before : scheduler.scheduleAtFixedRate(new LoanOverdueChecker(), 1, 24 * 60, TimeUnit.MINUTES);
    agter  : scheduler.scheduleAtFixedRate(new LoanOverdueChecker(), 1, (long )24 * 60, TimeUnit.MINUTES);

+++ src/main/java/lms/presentation/LibraryApp.java
@@ -43,7 +43,7 @@ public class LibraryApp {
 
 	public static void main(String[] args) {
 	 
-		scheduler.scheduleAtFixedRate(new LoanOverdueChecker(), 1, 24 * 60, TimeUnit.MINUTES);
+		scheduler.scheduleAtFixedRate(new LoanOverdueChecker(), 1, (long )24 * 60, TimeUnit.MINUTES);
 
 		 
 		UserRepository userRepo = StaticUserRepository.getInstance();
